requirements.txt
altgraph>=0.17.4
anyio>=4.8.0
certifi>=2024.12.14
colorama>=0.4.6
h11>=0.14.0
httpcore>=1.0.7
httpx>=0.28.1
idna>=3.10
packaging>=24.2
pefile>=2023.2.7
pillow>=11.1.0
pyinstaller>=6.11.1
pyinstaller-hooks-contrib>=2024.11
pywin32-ctypes>=0.2.3
pyzstd>=0.16.2
setuptools>=75.8.0
sniffio>=1.3.1
tqdm>=4.67.1
typing_extensions>=4.12.2